Finished the order screen, could probably use a little more styling
