+++ DroneCones/DroneConesApp/static/DroneConesApp/css/orderpage.css
@@ -101,3 +101,28 @@
     list-style-type: none;
     font-size: 20px;
 }
+
+.button{
+    padding: .25em;
+    cursor:pointer;
+    font-size: 14px;
+    border: 2px solid #000;
+    border-radius: 12px;
+    margin-bottom: 20px;
+    background-color: #fccc19
+}
+
+.order-entry{
+    display: inline-grid;
+    justify-items: center;
+    position: relative;
+    text-decoration: none;
+    color: black;
+    padding: .5em;
+    margin: 1em 1em 1em 1em;
+    border-radius: 8px;
+}
+
+.order-entry span{
+    font-size: 20px;
+}
