--- DroneCones/DroneConesApp/templates/DroneConesApp/OrderFlow/OrderParts/item.html
@@ -1,7 +1,0 @@
-
-{% load static %}
-<link rel="stylesheet" href="{% static 'DroneConesApp/css/orderpage.css' %}">
-<div  class="item">
-<span class="dot">Ice Cream image</span>
-<p> Ice cream </p>   
-</div>
