+++ DroneCones/DroneConesApp/templates/DroneConesApp/Orders/checkout.html
@@ -1,0 +1,15 @@
+{% extends "DroneConesApp/misc/base.html" %}
+
+{% block title %}Checkout{% endblock %}
+
+{% load static %}
+{% block content %}
+<link rel="stylesheet" href="{% static 'DroneConesApp/css/orderpage.css' %}">
+<div class="content-container">
+    <h1 style="text-align: center; font-size: 30px">Checkout</h1>
+    <hr style="margin-top: -20px; margin-bottom: 30px">
+    <div class="sub-content-container">
+
+    </div>
+</div>
+{% endblock %}
